split CRVReco
