+++ CRVAnalysis/src/SConscript
@@ -25,6 +25,7 @@ mainlib = helper.make_mainlib ( [ 'CLHEP',
                                   'mu2e_GeometryService',
                                   'mu2e_CosmicRayShieldGeom',
                                   'mu2e_CRVResponse',
+                                  'mu2e_CRVReco',
                                   'mu2e_GlobalConstantsService_GlobalConstantsService_service',
                                   'mu2e_GlobalConstantsService',
                                   'mu2e_ConfigTools',
