+++ CRVReco/src/SConscript
@@ -1,0 +1,69 @@
+#!/usr/bin/env python
+#
+# Original author Rob Kutschke.
+#
+
+import os, re
+Import('env')
+
+Import('mu2e_helper')
+
+helper=mu2e_helper(env)
+
+rootlibs = env['ROOTLIBS']
+
+mainlib = helper.make_mainlib ( [ 'CLHEP',
+                                  'mu2e_Mu2eUtilities',
+                                  'mu2e_RecoDataProducts',
+                                  'mu2e_DataProducts',
+                                  'mu2e_GeomPrimitives',
+                                  'mu2e_Mu2eInterfaces',
+                                  'art_Framework_Principal',
+                                  'art_Persistency_Provenance',
+                                  'art_Persistency_Common',
+                                  'art_Utilities',
+                                  'canvas',
+                                  'cetlib',
+                                  'cetlib_except',
+                                  rootlibs]
+                              )
+
+helper.make_plugins( [ mainlib,
+                       'mu2e_SeedService_SeedService_service',
+                       'mu2e_Mu2eUtilities',
+                       'mu2e_GeometryService',
+                       'mu2e_ConditionsService',
+                       'mu2e_DAQConditions',
+                       'mu2e_GlobalConstantsService_GlobalConstantsService_service',
+                       'mu2e_GlobalConstantsService',
+                       'mu2e_RecoDataProducts',
+                       'mu2e_CosmicRayShieldGeom',
+                       'mu2e_Mu2eInterfaces',
+                       'mu2e_ConfigTools',
+                       'mu2e_DataProducts',
+                       'mu2e_DbTables',
+                       'art_Framework_Core',
+                       'art_Framework_Principal',
+                       'art_Framework_Services_Registry',
+                       'art_root_io_tfile_support',
+                       'art_root_io_TFileService',
+                       'art_Framework_Services_Optional_RandomNumberGenerator_service',
+                       'art_Persistency_Common',
+                       'art_Persistency_Provenance',
+                       'art_Utilities',
+                       'canvas',
+                       'MF_MessageLogger',
+                       'fhiclcpp',
+                       'fhiclcpp_types',
+                       'tbb',
+                       'cetlib',
+                       'cetlib_except',
+                       'CLHEP',
+                       rootlibs,
+                       'boost_filesystem',
+                       ] )
+
+# this tells emacs to view this file in python mode.
+# Local Variables:
+# mode:python
+# End:
